Add skeleton for CoAP client

+++ src/coap/atlas_coap_client.h
@@ -14,6 +14,7 @@ size_t resp_payload_len);
  * @param[in] cb Response callback
  * @return status
  */
-atlas_status_t atlas_coap_client_request(const char *uri, int port, const uint8_t *req_payload, size_t req_payload_len, atlas_coap_client_cb_t cb);
+atlas_status_t atlas_coap_client_request(const char *uri, uint16_t port,
+const uint8_t *req_payload, size_t req_payload_len, atlas_coap_client_cb_t cb);
 
 #endif /* __ATLAS_COAP_CLIENT_H__ */
